pyproject.toml
[tool.poetry]
name = "data-analyst"
version = "0.1.0"
description = "Data Analyst Agent - Automated cost center analysis with anomaly detection and actionable insights"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "data_analyst_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^1.16.0"
a2a-sdk = "^0.3.8"
google-cloud-aiplatform = "^1.120.0"
pandas = "^2.3.3"
numpy = "^2.2.6"
pyodbc = "^5.2.0"
pyyaml = "^6.0.3"
tableauhyperapi = "^0.0.23135"
statsmodels = "^0.14.5"
scipy = "^1.15.3"
scikit-learn = "^1.7.2"
python-dateutil = "^2.9.0"
ruptures = "^1.1.10"
matplotlib = "^3.10.7"
python-dotenv = "^1.1.1"
absl-py = "^2.2.1"
fastapi = "^0.118.2"
uvicorn = "^0.37.0"
sqlalchemy = "^2.0.43"
pydantic = "^2.12.0"
pydantic-settings = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

